temp/u-boot-pinephone: drop u-boot in favor of Tow-Boot
Just as we did with the PPP already, we're moving the PP to requiring
Tow-Boot to be installed. This allows us to drop our own u-boot package
and use standard UEFI-based booting. We'll move to UEFI later but we can
already drop our own u-boot and use Tow-Boot in combination with
extlinux.conf